Shifted helper functions to helpers file.
